+++ packages/graphql/src/translate/translate-top-level-cypher.ts
@@ -158,16 +158,18 @@ export function translateTopLevelCypher({
     const apocParamsStr = `{${apocParams.strs.length ? `${apocParams.strs.join(", ")}` : ""}}`;
 
     if (type === "Query") {
-        const isArray = field.typeMeta.array;
-        const expectMultipleValues = !field.isScalar && !field.isEnum && isArray;
-
-        if (expectMultipleValues) {
-            cypherStrs.push(`WITH apoc.cypher.runFirstColumnMany("${statement}", ${apocParamsStr}) as x`);
+        if (field.columnName) {
+            const experimentalCypherStatement = createCypherDirectiveSubquery({
+                field,
+            });
+            cypherStrs.push(...experimentalCypherStatement);
         } else {
-            cypherStrs.push(`WITH apoc.cypher.runFirstColumnSingle("${statement}", ${apocParamsStr}) as x`);
+            const legacyCypherStatement = createCypherDirectiveApocProcedure({
+                field,
+                apocParams: apocParamsStr,
+            });
+            cypherStrs.push(...legacyCypherStatement);
         }
-
-        cypherStrs.push("UNWIND x as this\nWITH this");
     } else {
         cypherStrs.push(`
             CALL apoc.cypher.doIt("${statement}", ${apocParamsStr}) YIELD value
@@ -203,4 +205,39 @@ export function translateTopLevelCypher({
         }
         return [cypherStrs.join("\n"), params];
     }).build();
+}
+
+function createCypherDirectiveApocProcedure({
+    field,
+    apocParams,
+}: {
+    field: CypherField;
+    apocParams: string;
+}): string[] {
+    const isArray = field.typeMeta.array;
+    const expectMultipleValues = !field.isScalar && !field.isEnum && isArray;
+    const cypherStrs: string[] = [];
+
+    if (expectMultipleValues) {
+        cypherStrs.push(`WITH apoc.cypher.runFirstColumnMany("${field.statement}", ${apocParams}) as x`);
+    } else {
+        cypherStrs.push(`WITH apoc.cypher.runFirstColumnSingle("${field.statement}", ${apocParams}) as x`);
+    }
+
+    cypherStrs.push("UNWIND x as this\nWITH this");
+    return cypherStrs;
+}
+
+function createCypherDirectiveSubquery({ field }: { field: CypherField }): string[] {
+    const cypherStrs: string[] = [];
+    cypherStrs.push("CALL {", field.statement, "}");
+
+    if (field.columnName) {
+        if (field.isScalar || field.isEnum) {
+            cypherStrs.push(`UNWIND ${field.columnName} as this`);
+        } else {
+            cypherStrs.push(`WITH ${field.columnName} as this`);
+        }
+    }
+    return cypherStrs;
 }