+++ packages/graphql/tests/tck/aggregations/where/node/time.test.ts
@@ -68,7 +68,7 @@ describe("Cypher Aggregations where node with Time", () => {
                 RETURN any(var2 IN collect(this1.someTime) WHERE var2 = $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
@@ -80,8 +80,7 @@ describe("Cypher Aggregations where node with Time", () => {
                     \\"second\\": 0,
                     \\"nanosecond\\": 0,
                     \\"timeZoneOffsetSeconds\\": 0
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -108,7 +107,7 @@ describe("Cypher Aggregations where node with Time", () => {
                 RETURN any(var2 IN collect(this1.someTimeAlias) WHERE var2 = $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
@@ -120,8 +119,7 @@ describe("Cypher Aggregations where node with Time", () => {
                     \\"second\\": 0,
                     \\"nanosecond\\": 0,
                     \\"timeZoneOffsetSeconds\\": 0
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -148,7 +146,7 @@ describe("Cypher Aggregations where node with Time", () => {
                 RETURN any(var2 IN collect(this1.someTime) WHERE var2 > $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
@@ -160,8 +158,7 @@ describe("Cypher Aggregations where node with Time", () => {
                     \\"second\\": 0,
                     \\"nanosecond\\": 0,
                     \\"timeZoneOffsetSeconds\\": 0
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -188,7 +185,7 @@ describe("Cypher Aggregations where node with Time", () => {
                 RETURN any(var2 IN collect(this1.someTime) WHERE var2 >= $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
@@ -200,8 +197,7 @@ describe("Cypher Aggregations where node with Time", () => {
                     \\"second\\": 0,
                     \\"nanosecond\\": 0,
                     \\"timeZoneOffsetSeconds\\": 0
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -228,7 +224,7 @@ describe("Cypher Aggregations where node with Time", () => {
                 RETURN any(var2 IN collect(this1.someTime) WHERE var2 < $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
@@ -240,8 +236,7 @@ describe("Cypher Aggregations where node with Time", () => {
                     \\"second\\": 0,
                     \\"nanosecond\\": 0,
                     \\"timeZoneOffsetSeconds\\": 0
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -268,7 +263,7 @@ describe("Cypher Aggregations where node with Time", () => {
                 RETURN any(var2 IN collect(this1.someTime) WHERE var2 <= $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
@@ -280,8 +275,7 @@ describe("Cypher Aggregations where node with Time", () => {
                     \\"second\\": 0,
                     \\"nanosecond\\": 0,
                     \\"timeZoneOffsetSeconds\\": 0
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -308,7 +302,7 @@ describe("Cypher Aggregations where node with Time", () => {
                 RETURN min(this1.someTime) = $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -320,8 +314,7 @@ describe("Cypher Aggregations where node with Time", () => {
                     \\"second\\": 0,
                     \\"nanosecond\\": 0,
                     \\"timeZoneOffsetSeconds\\": 0
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -348,7 +341,7 @@ describe("Cypher Aggregations where node with Time", () => {
                 RETURN min(this1.someTime) > $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -360,8 +353,7 @@ describe("Cypher Aggregations where node with Time", () => {
                     \\"second\\": 0,
                     \\"nanosecond\\": 0,
                     \\"timeZoneOffsetSeconds\\": 0
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -388,7 +380,7 @@ describe("Cypher Aggregations where node with Time", () => {
                 RETURN min(this1.someTime) >= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -400,8 +392,7 @@ describe("Cypher Aggregations where node with Time", () => {
                     \\"second\\": 0,
                     \\"nanosecond\\": 0,
                     \\"timeZoneOffsetSeconds\\": 0
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -428,7 +419,7 @@ describe("Cypher Aggregations where node with Time", () => {
                 RETURN min(this1.someTime) < $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -440,8 +431,7 @@ describe("Cypher Aggregations where node with Time", () => {
                     \\"second\\": 0,
                     \\"nanosecond\\": 0,
                     \\"timeZoneOffsetSeconds\\": 0
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -468,7 +458,7 @@ describe("Cypher Aggregations where node with Time", () => {
                 RETURN min(this1.someTime) <= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -480,8 +470,7 @@ describe("Cypher Aggregations where node with Time", () => {
                     \\"second\\": 0,
                     \\"nanosecond\\": 0,
                     \\"timeZoneOffsetSeconds\\": 0
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -508,7 +497,7 @@ describe("Cypher Aggregations where node with Time", () => {
                 RETURN max(this1.someTime) = $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -520,8 +509,7 @@ describe("Cypher Aggregations where node with Time", () => {
                     \\"second\\": 0,
                     \\"nanosecond\\": 0,
                     \\"timeZoneOffsetSeconds\\": 0
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -548,7 +536,7 @@ describe("Cypher Aggregations where node with Time", () => {
                 RETURN max(this1.someTime) > $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -560,8 +548,7 @@ describe("Cypher Aggregations where node with Time", () => {
                     \\"second\\": 0,
                     \\"nanosecond\\": 0,
                     \\"timeZoneOffsetSeconds\\": 0
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -588,7 +575,7 @@ describe("Cypher Aggregations where node with Time", () => {
                 RETURN max(this1.someTime) >= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -600,8 +587,7 @@ describe("Cypher Aggregations where node with Time", () => {
                     \\"second\\": 0,
                     \\"nanosecond\\": 0,
                     \\"timeZoneOffsetSeconds\\": 0
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -628,7 +614,7 @@ describe("Cypher Aggregations where node with Time", () => {
                 RETURN max(this1.someTime) < $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -640,8 +626,7 @@ describe("Cypher Aggregations where node with Time", () => {
                     \\"second\\": 0,
                     \\"nanosecond\\": 0,
                     \\"timeZoneOffsetSeconds\\": 0
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -668,7 +653,7 @@ describe("Cypher Aggregations where node with Time", () => {
                 RETURN max(this1.someTime) <= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -680,8 +665,7 @@ describe("Cypher Aggregations where node with Time", () => {
                     \\"second\\": 0,
                     \\"nanosecond\\": 0,
                     \\"timeZoneOffsetSeconds\\": 0
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });